requirements.txt
aiogram==3.18.0
Flask==3.1.0
Pillow==11.1.0
pillow_heif==0.21.0
pytesseract==0.3.13
python-dotenv==1.0.1
pyzbar==0.1.9
Requests==2.32.3
zxing==1.0.3
